move db to main app
