+++ PosContract/build.gradle
@@ -27,8 +27,4 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    api project(':database')
 }
-
-//sourceCompatibility = "1.7"
-//targetCompatibility = "1.7"
