+++ PosContract/src/main/java/com/paystack/poscontract/models/Transaction.kt
@@ -1,4 +1,4 @@
-package com.example.database
+package com.paystack.poscontract.models
 
 data class Transaction(
     var paystackReference: String = "",
