+++ PosContract/src/main/java/com/paystack/poscontract/models/TransactionType.kt
@@ -1,5 +1,5 @@
 package com.paystack.poscontract.models
 
 enum class TransactionType {
-    PURCHASE, CASH_ADVAANCE
+    PURCHASE, CASH_ADVANCE
 }
