+++ app/build.gradle
@@ -56,10 +56,12 @@ dependencies {
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     implementation 'androidx.navigation:navigation-fragment-ktx:2.1.0'
     implementation 'androidx.navigation:navigation-ui-ktx:2.1.0'
+    implementation 'androidx.room:room-runtime:2.2.5'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
     implementation project(':PosContract')
+    annotationProcessor 'androidx.room:room-compiler:2.2.5'
     nexgoImplementation project(':hardware-nexgo')
     huaweiiImplementation project(':hardware-huaweii')
     epmsImplementation project(':processors-epms')
