+++ app/src/main/java/com/paystack/architecturedemo/database/AppDatabase.kt
@@ -1,9 +1,9 @@
-package com.example.database
+package com.paystack.architecturedemo.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
 
-@Database(entities = [Transaction::class], version = 1)
+@Database(entities = [TransactionEntity::class], version = 1)
 abstract class AppDatabase : RoomDatabase() {
     abstract fun transactionDataDao(): TransactionDao
 }
