--- database/src/main/AndroidManifest.xml
@@ -1,5 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.database">
-
-    /
-</manifest>
