+++ hardware-nexgo/src/main/java/com/paystack/hardware-nexgo/NexgoApp.kt
@@ -3,6 +3,7 @@ package com.paystack.hardware_nexgo
 import android.content.Context
 import com.paystack.poscontract.Hardware
 import com.paystack.poscontract.models.ReceiptType
+import com.paystack.poscontract.models.Transaction
 
 class NexgoApp : Hardware {
 
@@ -14,7 +15,12 @@ class NexgoApp : Hardware {
         TODO("Not yet implemented")
     }
 
-    override fun printReceipt(receiptType: ReceiptType, isReprint: Boolean) {
+    override fun printReceipt(
+        receiptType: ReceiptType,
+        isReprint: Boolean,
+        transaction: Transaction
+    ) {
         TODO("Not yet implemented")
     }
+
 }
