+++ processors-tams/src/main/java/com/paystack/processors_tams/TamsApp.kt
@@ -1,7 +1,7 @@
 package com.paystack.processors_tams
 
-import com.example.database.Transaction
 import com.paystack.poscontract.Processor
+import com.paystack.poscontract.models.Transaction
 import com.paystack.poscontract.models.TransactionType
 
 class TamsApp : Processor {
