+++ settings.gradle
@@ -1,4 +1,3 @@
-include ':database'
 include ':processors-tams'
 include ':processors-epms'
 include ':hardware-huaweii'
